consolidate migrations

+++ db/migrations/00009_create_eth_storage_cids_table.sql
@@ -1,7 +1,7 @@
 -- +goose Up
 CREATE TABLE eth.storage_cids (
-  id                    SERIAL PRIMARY KEY,
-  state_id              INTEGER NOT NULL REFERENCES eth.state_cids (id) ON DELETE CASCADE DEFERRABLE INITIALLY DEFERRED,
+  id                    BIGSERIAL PRIMARY KEY,
+  state_id              BIGINT NOT NULL REFERENCES eth.state_cids (id) ON DELETE CASCADE DEFERRABLE INITIALLY DEFERRED,
   storage_leaf_key      VARCHAR(66),
   cid                   TEXT NOT NULL,
   mh_key                TEXT NOT NULL REFERENCES public.blocks (key) ON DELETE CASCADE DEFERRABLE INITIALLY DEFERRED,
